Improve command line interface for subtitles

+++ src/subtitles.py
@@ -1,9 +1,9 @@
 from faster_whisper.transcribe import Segment, Word
-import re
 from transcribe import transcribe, TranscribeOptions
 from filters import compile_filters, filter_transcription
 import ffmpeg
-import os, pathlib
+import os, pathlib, argparse
+from cli import confirm
 from dataclasses import dataclass
 
 @dataclass
@@ -128,8 +128,11 @@ def _probability_to_alpha(probability: float) -> int:
 
 def add_subtitles_to_video(video_file: str, subtitles_script: str, output_file: str):
   if os.path.isfile(output_file):
-    # TODO: Confirm before continuing
-    os.remove(output_file)
+    if confirm("Output file already exists. Overwrite it?", default=False):
+      os.remove(output_file)
+    else:
+      print("Canceled")
+      exit(1)
   av = ffmpeg.input(video_file)
   subtitles = ffmpeg.input('pipe:')
   stream = ffmpeg.output(av, subtitles, output_file, c="copy", loglevel="error")
@@ -138,6 +141,39 @@ def add_subtitles_to_video(video_file: str, subtitles_script: str, output_file: 
   process.stdin.close()
   if process.wait() != 0:
     print("Error from ffmpeg")
+
+def _parse_arguments() -> argparse.Namespace:
+  parser = argparse.ArgumentParser(
+    prog='subtitles',
+    description='A command-line tool for adding subtitles to automute-filtered videos.'
+  )
+  parser.add_argument('input',
+                      help='Audio or video file to add subtitles to.')
+  parser.add_argument('-o', '--output',
+                      help='Name of the output file. (Default: <input file>-subtitled.mkv)')
+  parser.add_argument('-w', '--filter-word', default=[], action='append',
+                      help='A word to filter out. Treated as a case-insensitive regular expression. Can be specified multiple times.')
+  parser.add_argument('-f', '--filter-file', default=[], action='append',
+                      help="A file of words to filter out. Each line is treated as a case-insentive regular expression.")
+  parser.add_argument('-e', '--encipher-words', default=False, action='store_true',
+                      help='Before applying filters, enchiper transcribed words by replacing each letter with the one immediately after it in the alphabet ' +
+                           '(looping around at the end; i.e., caesar cipher with a shift of 1). Use this if you need to filter out profanity but would prefer ' +
+                           'not to read profanity when specifying your filters.')
+  parser.add_argument('--whisper-model', default='large-v3',
+                      help='The faster_whisper model to use for audio transcription. Can be a model name (tiny, tiny.en, base, base.en, small, small.en, ' +
+                           'distil-small.en, medium, medium.en, distil-medium.en, large, large-v1, large-v2, distil-large-v2, large-v3, or distil-large-v3), ' +
+                           'a CTranslate2-converted model ID from Hugging Face, or a path to a local model. See faster-whisper docs. (Default: large-v3)')
+  parser.add_argument('--whisper-compute-type', default='auto', choices=['default','auto','int8','int8_float16','int8_bfloat16','int8_float32','int16','float16','bfloat16','float32'],
+                      help='The compute type to use when loading the Whisper model. \'default\' explicitly uses the same quantization that the model is already using. ' + 
+                           '\'auto\' selects the fasted option that is supported by the device used. (Default: auto)')
+  parser.add_argument('--whisper-device', default='auto', choices=['auto','cpu','cuda'],
+                      help='The compute device to use when running the Whisper model. (Default: auto)')
+  # parser.add_argument('--whisper-silence-ms', default=-1, type=int,
+  #                     help='The minimum duration in milliseconds for an audio segment with no detected speech to be skipped during transcription. -1 to disable. ' +
+  #                          '(Default: -1)')
+  parser.add_argument('--ignore-cached-transcriptions', default=False, action='store_true',
+                      help='Ignore any cached transcriptions.')
+  return parser.parse_args()
 
 def _get_filtered_video_path(input_file: str) -> str:
   """Creates an output file path from an input file path."""
@@ -150,20 +186,38 @@ def _get_output_file_path(input_file: str) -> str:
   return str(input_path.with_stem(input_path.stem + "-subtitled").with_suffix(".mkv"))
 
 def main():
-  filter_list = compile_filters([], ['default_wordlist_en.txt'])
+  args = _parse_arguments()
 
-  video_file = "./media/missile.mp4"
-  segments = transcribe(video_file, TranscribeOptions(model="large-v3"))
-  segments, matches = filter_transcription(segments, filter_list, '[__]', True)
+  input_file = args.input
+  filtered_file = _get_filtered_video_path(input_file)
+  output_file = args.output if args.output is not None else _get_output_file_path(input_file)
+  
+  filters = compile_filters(args.filter_word, args.filter_file)
+
+  segments = transcribe(
+    input_file,
+    TranscribeOptions(
+      model=args.whisper_model,
+      device=args.whisper_device,
+      compute_type=args.whisper_compute_type,
+      condition_on_previous_text='distil' not in args.whisper_model, # Distil models seem prone to repeating themselves
+      # hotwords=[decipher(word) if args.encipher_words else word for f in filters for word in [f.pattern[2:-2]]],
+      # vad_filter=args.whisper_silence_ms >= 0,
+      # vad_parameters=dict(
+      #   min_silence_duration_ms=args.whisper_silence_ms
+      # ) if args.whisper_silence_ms >= 0 else dict()
+    ),
+    ignore_cache=args.ignore_cached_transcriptions,
+  )
+
+  segments, matches = filter_transcription(segments, filters, '[__]', args.encipher_words)
   print(f"Found {matches} matches for filters")
+
   subtitles = layout_subtitles(segments)
   script = create_subtitles_script(subtitles)
 
-  output_file = _get_output_file_path(video_file)
-  add_subtitles_to_video(_get_filtered_video_path(video_file), script, output_file)
-  print("Added subtitles to video")
-
-  # TODO: https://www.bannerbear.com/blog/how-to-add-subtitles-to-a-video-file-using-ffmpeg/#hard-subtitles-vs-soft-subtitles
+  add_subtitles_to_video(filtered_file, script, output_file)
+  print("Done")
 
 if __name__ == "__main__":
   main()
